[feat #71] 웰니스 정보 썸네일 이미지 AWS S3 저장 위치 변경 - 한국관광공사 API 제공 이미지를 저장하기 위해 새롭게 생성한 버킷 삭제, 기존 버킷 내에서 폴더로 저장 위치를 구분하도록 변경 - 이미지 S3 저장 로직을 하나의 파일(S3Service) 내에서 관리하도록 변경
